Disable the recorder feature by default
In order to be really fun, I must add a new window where it will be
possible to load a record exactly like for the savegame. A button should
be added to the main screen in order to access to the records.

I suggets to add this feature with the next version.