Add the option for chaning the render quality
It toggles between 'nearest' (faster) and 'best' (slower but more
prettier). It's enabled only when the game runs in fullscreen.

The option can be passed by command line argument `--render-quality on`,
by settings.json file with a `"renderquality": true` entry or simply
with the main settings screen in the game.