Fix mouse sprites and hovers when an update is doing
Otherwise it's possible to have de desync between the current state of
the last update and the last time that the mouse was handled. It works
much better now because it's no longer possible to have for example a
red cross on the ground when a tomato has finished to grow. The mouse
state is immediatly fixed by the update.