Remove buggy suspend stuff
It doesn't make sense here. Not sure why the old
code needs that. Anyway, it's just wrong because.
Maybe the old code needs that because the sound
was played independently of the video.

This commit prevents that the music to continue
on the win or lost screen when the video has been
skipped by space or esc.